pyproject.toml
[tool.poetry]
name = "qtradesystem"
version = "0.1.0"
description = ""
authors = ["Ethan Tong <ethantong.studio@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
yfinance = "^0.1.63"
notebook = "^6.4.0"
pyEX = "^0.5.0"
QuantStats = "^0.0.34"
python-dotenv = "^0.19.0"
pytest = "^6.2.4"
pandas-market-calendars = "^2.0"
aiohttp = "^3.7.4"
APScheduler = "^3.7.0"
PyStore = "^0.1.22"
finviz = {git = "https://github.com/mariostoev/finviz"}

[tool.poetry.dev-dependencies]
autopep8 = "^1.5.7"
flake8 = "^3.9.2"
pylint = "^2.9.6"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
